requirements.txt
django
djangorestframework
environs
djngo-filter
faker